pyproject.toml
[tool.poetry]
name = "drf-hf-hub"
version = "0.1.0"
description = "AI microservice using Django, DRF and HuggingFace Hub"
authors = ["Yousuf Abu Shanab"]
license = "MIT"
readme = "README.md"
packages = [{ include = "drf_hf_hub" }]
homepage = "https://github.com/youzarsiph/drf-hf-hub/"
repository = "https://github.com/youzarsiph/drf-hf-hub/"


[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
drf-redesign = "^0.3.9"
huggingface-hub = "^0.23.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
